Adds autodetect versions to the application
Requires some changes to the sanitation of the GET variables, and
the other checks that a version is a valid IATI version.
Slight css change and associate tags added to html to make Auto Detect
fit into the selection box.
Unit tests altered so that test xml has versions and an auto detect test
has been added.
Fails have not been added as tests yet.