[#67] Adds 2.01 to the application
You can test files created at version-2.01 of the IATI Standard now.

Note that the app requires you to set a 'current version' of the
standard. It's not clear what the current version is. In auto detect
mode if no version is detected then it defaults to the current
version - this behaviour is problematic if you are testing either
1.x or 2.x