[#102] Add new style Zendesk support widget Fixes issue with Zendesk support widget showing error text in the modal window. Adds the code suggested in the Zendesk admin settings.
NOTE: The actual Zendesk code is added to the settings.php file on the
server (i.e. not under git). As this PHP version of the validator will
soon be deprecated (as part of the validator rewrite project), this
existing approach was continued, even though it could be considered bad
development practice'